moved load_config __main__ -> eatlocal
